ipcache: Remove metric for idempotent operations
This metric was being counted and regularly reported in high-scale
clusters. The error field of the metric reported "idempotent_operation",
but looking at the codepath we can see that there is no error and this
is an expected condition. We now know for sure that users will hit this
path in production, no need to count it.

Signed-off-by: Joe Stringer <joe@cilium.io>